pyproject.toml
[project]
name = "cabinets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
    "selenium-requests>=2.0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
